同じファイル名でも適切に生成される

+++ panyubi/logic/process.py
@@ -1,3 +1,4 @@
+from genericpath import isfile
 from glob import glob
 import re
 import cv2
@@ -32,14 +33,15 @@ def is_frame_to_use(bf, detector, img_path, pre_img_path, pre_des):
     img_num = int(img_num)
     return img_num, score_match, des
 
-def clear_files(img_directory_path):
-    search_path = img_directory_path + "/*"
-    remove_path_li = glob(search_path)
+def clear_files(directory_path):
+    search_path = directory_path + "/**"
+    remove_path_li = glob(search_path, recursive=True)
+    print(remove_path_li)
     for remove_path in remove_path_li:
-        os.remove(remove_path) 
+        if os.path.isfile(remove_path) :
+            os.remove(remove_path)
 
 def pre_process(img_directory_path):
-    clear_files(img_directory_path)
     search_path = img_directory_path + "/*.png"
     img_path_li = sorted(glob(search_path), key=natural_keys)
     EXT = ".png"
